requirements.txt
asgiref==3.8.1
Django==5.1
django-tinymce==4.1.0
pillow==10.4.0
sqlparse==0.5.1
